requirements.txt
PyMuPDF>=1.26.6
Pillow>=12.0.0
anthropic>=0.74.1
python-dotenv>=1.2.1
